KNGS_Processes_Website/requirements.txt
azure-identity==1.15.0
msgraph-sdk==1.0.0
flask==2.3.3
flask-cors==4.0.0
configparser==6.0.0
requests==2.31.0 